+++ internal/provider/provider.go
@@ -35,6 +35,7 @@ func New() *schema.Provider {
 		ResourcesMap: map[string]*schema.Resource{
 			"bunny_pullzone": resourcePullZone(),
 			"bunny_edgerule": resourceEdgeRule(),
+			"bunny_hostname": resourceHostname(),
 		},
 		ConfigureContextFunc: newProvider,
 	}